Merge to mainline: Release 87

+++ ESP32/src/ForceCurve.cpp
@@ -33,9 +33,9 @@ float ForceCurve_Interpolated::EvalForceCubicSpline(const DAP_config_st* config_
   yOrig[5] = config_st->payLoadPedalConfig_.relativeForce_p100;
 
   //double dx = 1.0f;
-  double t = (splineSegment_fl32 - (float)splineSegment_u8);// / dx;
-  double y = (1 - t) * yOrig[splineSegment_u8] + t * yOrig[splineSegment_u8 + 1] + t * (1 - t) * (a * (1 - t) + b * t);
-
+  float t = (splineSegment_fl32 - (float)splineSegment_u8);// / dx;
+  float y = (1 - t) * yOrig[splineSegment_u8] + t * yOrig[splineSegment_u8 + 1] + t * (1 - t) * (a * (1 - t) + b * t);
+  
   if (calc_st->Force_Range> 0)
   {
       y = calc_st->Force_Min + y / 100.0f * calc_st->Force_Range;
@@ -79,11 +79,11 @@ float ForceCurve_Interpolated::EvalForceGradientCubicSpline(const DAP_config_st*
 
 
 
-  double Delta_x_orig = 100; // total horizontal range [0,100]
-  double dx = Delta_x_orig / NUMBER_OF_SPLINE_SEGMENTS; // spline segment horizontal range
-  double t = (splineSegment_fl32 - (float)splineSegment_u8); // relative position in spline segment [0, 1]
-  double dy = yOrig[splineSegment_u8 + 1] - yOrig[splineSegment_u8]; // spline segment vertical range
-  double y_prime = 0;
+  float Delta_x_orig = 100; // total horizontal range [0,100]
+  float dx = Delta_x_orig / NUMBER_OF_SPLINE_SEGMENTS; // spline segment horizontal range
+  float t = (splineSegment_fl32 - (float)splineSegment_u8); // relative position in spline segment [0, 1]
+  float dy = yOrig[splineSegment_u8 + 1] - yOrig[splineSegment_u8]; // spline segment vertical range
+  float y_prime = 0;
   if (fabs(dx) > 0)
   {
       y_prime = dy / dx + (1 - 2 * t) * (a * (1 - t) + b * t) / dx + t * (1 - t) * (b - a) / dx;
@@ -92,8 +92,8 @@ float ForceCurve_Interpolated::EvalForceGradientCubicSpline(const DAP_config_st*
   // --> conversion of the gradient to the proper axis scaling is performed
   if (normalized_b == false)
   {
-    double d_y_scale = calc_st->Force_Range / 100.0;
-    double d_x_scale=0;
+    float d_y_scale = calc_st->Force_Range / 100.0;
+    float d_x_scale=0;
     if (fabs(calc_st->stepperPosRange) > 0.01)
     {
         d_x_scale = 100.0 / calc_st->stepperPosRange;
